+++ backend/src/utils/import-exercise.ts
@@ -1,6 +1,6 @@
 import type { ExerciseKeys, StateExport } from 'digital-fuesim-manv-shared';
 import { ReducerError } from 'digital-fuesim-manv-shared';
-import type { ActiveExercise } from '../exercise/exercise-wrapper.js';
+import type { ActiveExercise } from '../exercise/active-exercise.js';
 import type { HttpResponse } from '../exercise/http-handler/utils.js';
 import { ExerciseFactory } from '../exercise/exercise-factory.js';
 import { ValidationErrorWrapper } from './validation-error-wrapper.js';
