+++ backend/test/exercise-service.spec.ts
@@ -1,7 +1,7 @@
 import { jest } from '@jest/globals';
 import { uuid } from 'digital-fuesim-manv-shared';
 import { createExercise, createTestEnvironment } from '../test/utils.js';
-import { ActiveExercise } from '../src/exercise/exercise-wrapper.js';
+import { ActiveExercise } from '../src/exercise/active-exercise.js';
 
 describe('Exercise-Service', () => {
     const environment = createTestEnvironment();
