+++ backend/test/websocket-join-exercise.spec.ts
@@ -1,7 +1,7 @@
 import assert from 'node:assert';
 import { jest } from '@jest/globals';
 import { generateDummyPatient, sleep } from 'digital-fuesim-manv-shared';
-import { ActiveExercise } from '../src/exercise/exercise-wrapper.js';
+import { ActiveExercise } from '../src/exercise/active-exercise.js';
 import { createExercise, createTestEnvironment } from './utils.js';
 
 describe('join exercise', () => {
@@ -50,7 +50,8 @@ describe('join exercise', () => {
 
     it('ignores clients joining other exercises', async () => {
         const firstExerciseKey = (await createExercise(environment)).trainerKey;
-        const secondExerciseKey = (await createExercise(environment)).trainerKey;
+        const secondExerciseKey = (await createExercise(environment))
+            .trainerKey;
 
         await environment.withWebsocket(async (firstClientSocket) => {
             const firstClientName = 'someRandomName';
